Group EndToEnd tests into a single file
