+++ EssentialFeedEndToEndTests/EssentialFeedEndToEndTests.swift
@@ -1,17 +1,17 @@
 import XCTest
 import EssentialFeed
 
-class RemoteFeedLoaderURLSessionHTTPClientTests: XCTestCase {
+class EssentialFeedEndToEndTests: XCTestCase {
 
-    func testRemoteFeedLoaderAndURLSessionHTTPClientReturnsCorrectFeedImages() {
+    func testRemoteFeedLoaderAndURLSessionHTTPClientReturnsCorrectFeedImages(file: StaticString = #filePath, line: UInt = #line) {
         let exp = expectation(description: "waiting for real request to complete")
 
         let url = URL(string: "https://essentialdeveloper.com/feed-case-study/test-api/feed")!
         let httpClient = URLSessionHTTPClient()
         let remoteFeedLoader = RemoteFeedLoader(url: url, client: httpClient)
 
-        testMemoryLeak(httpClient)
-        testMemoryLeak(remoteFeedLoader)
+        testMemoryLeak(httpClient, file: file, line: line)
+        testMemoryLeak(remoteFeedLoader, file: file, line: line)
 
         var capturedFeedImages = [FeedImage]()
         remoteFeedLoader.load { result in
@@ -29,6 +29,32 @@ class RemoteFeedLoaderURLSessionHTTPClientTests: XCTestCase {
         for index in 0...7 {
             XCTAssertEqual(capturedFeedImages[index], feedImage(at: index))
         }
+    }
+
+    func test_remoteImageLoader_and_URLSessionHTTPClient_deliversImageDataFromAPI(file: StaticString = #filePath, line: UInt = #line) {
+        let exp = expectation(description: "Wait for request to complete")
+        let testServerURL = URL(string: "https://essentialdeveloper.com/feed-case-study/test-api/feed/73A7F70C-75DA-4C2E-B5A3-EED40DC53AA6/image")!
+
+        let httpClient = URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
+        let imageLoader = RemoteImageLoader(client: httpClient)
+
+        testMemoryLeak(httpClient, file: file, line: line)
+        testMemoryLeak(imageLoader, file: file, line: line)
+
+        _ = imageLoader.load(from: testServerURL) { result in
+            switch (result) {
+            case .success(let data):
+                XCTAssertFalse(data.isEmpty, "Expected server to have returned some data")
+
+            case .failure:
+                XCTFail("Expected image data load to complete with success, instead got \(result)")
+
+            }
+
+            exp.fulfill()
+        }
+
+        wait(for: [exp], timeout: 30.0)
     }
 
     private func feedImage(at index: Int) -> FeedImage {
